Add export definition macros (#15)
* Add export definition macros

Windows requires extra attributes to export the symbols from a shared
library. Macros similar to the ones defined here are a typical way of
handling such attributes.

include/topotoolbox.h adds the TOPOTOOLBOX_API macro, which controls
which __declspec attribute is attached to exported API functions. The
TOPOTOOLBOX_BUILD flag controls whether the shared library is being
built, in which case the symbols must be exported, or used, in which
case they must be imported. An API function `has_topotoolbox` is
added. This function returns 1 (true), and can be used to test that
everything is linked correctly. As more functions are added to the
library, this functionality may become less useful.

src/topotoolbox.c defines the TOPOTOOLBOX_BUILD macro and the
`has_topotoolbox` function.

test/versioninfo.cpp now uses the `has_topotoolbox` function to check
before printing the version number.

test/CMakeLists.txt is modified to set the test PATH correctly on Windows.