Add continuous integration with GitHub actions (#16)
This creates a workflow that checks out the repository, then builds
and tests the repository with our CMake build system. The action is
run on Ubuntu with gcc and clang, macos with clang and Windows with
MSVC.

.github/workflows/ci.yaml contains the configuration for the CI
workflow. It is based loosely on the cmake-multi-platform starter
workflow provided by GitHub.