Implement sink filling using morphological reconstruction (#14)
topotoolbox fills sinks in digital elevation models using grayscale
morphological reconstruction. This algorithm repeatedly dilates a
"marker" image while constraining it to lie below a "mask" image. The
negative DEM is used as the mask while the marker is identical to the
negative DEM on the boundary pixels but negative infinity in its
interior. The morphological reconstruction iteratively fills in the
interior of the marker image, but does not descend into sinks of the
mask.

src/morphology/reconstruct.c implements the sequential reconstruction
algorithm presented by Vincent (1993). The algorithm repeatedly scans
in alternating directions through the DEM, updates the marker image
based on the values of the marker and mask in a neighborhood around
the given pixel and terminates when the marker stops changing. The
sequential reconstruction algorithm is not as efficient as a hybrid
algorithm presented by Vincent that combines sequential reconstruction
with a breadth-first search. However, the hybrid algorithm requires
managing a queue of pixels that remain to be updated, while the
sequential algorithm requires no memory allocation beyond the marker
and mask images.

src/morphology/reconstruct.h exposes the reconstruct function for use
elsewhere in topotoolbox. This private header is not, however, exposed
to users of the library.

src/fillsinks.c implements `fillsinks` using `reconstruct`. The input
and output digital elevation models should be passed in by the caller
as pointers to arrays of floats. The input DEM is temporarily
modified, but then reverted to its original state before the function
returns. This should be safe as long as concurrent accesses to the DEM
are avoided, but care should be taken in the future.

include/topotoolbox.h is modified to expose the `fillsinks` function
to users of the library.

test/fillsinks.cpp tests the `fillsinks` implementation by generating
100 random digital elevation models using a hash function based random
number generator and testing two properties of the filled DEM:

1. No pixel should be surrounded by pixels of a higher elevation.
2. No pixel should be below the original level of the DEM.

These tests do not completely constrain the sink filling algorithm,
however, and should be taken as a starting point for more thorough
testing.

src/CMakeLists.txt is modified to append fillsinks.c and
morphology/reconstruct.c to add_library, so that these are
compiled into the shared library.

test/CMakeLists.txt is modified to build the `fillsinks` executable
and add it to the test runner.