Initialize all pixels on the boundary in fillsinks (#22)
By using && in this conditional, this only initialized the
output/marker with pixels in the corners when we needed to initialize
all the boundary pixels. Switching the && to || ensures that all of
the boundary pixels are set equal to their value in the inverted DEM
before the reconstruction algorithm is run.