Contributing Guidelines (#38)
Adds a draft of contributing guidelines and moves the build instructions from the README to CONTRIBUTING.md.

The CONTRIBUTING.md includes a section for formatting and style, which provides a place to accumulate information on code style. We can always move that to a separate STYLE.md or similar if that section gets too large.

The README includes some links to other TopoToolbox projects and sites, some of which are not available yet. In particular the link for the libtopotoolbox documentation is broken, but it will be fixed shortly when #28 is resolved.