Implement excess topography using the fast marching method
src/excesstopography.c implements excess topography computation based
on solving the eikonal equation over the gridded DEM using the fast
marching method. The eikonal equation limits the maximum slope in
regions of the DEM where this slope is exceeded. In regions of low
slope, the resulting topography follows the original DEM. The return
value (float *excess) from the excesstopography function is not the
excess topography per se, but the constrained topography. The excess
topography can be computed by subtracting the constrained topography
from the original.

test/excesstopography.c implements property-based tests of the
excesstopography solver using a random DEM generated with the pcg4d
function used elsewhere. The only property currently tested is that the
excesstopography is constrained above by the DEM.