Implement 3D excess topography with the fast marching method
This implements a modification of the constrained fast marching method
to estimate excess topography with three-dimensional lithological
variability. The lithstack array is a (nlayers x nrows x ncols) array
containing the elevations of the top surfaces of each layer at each
grid cell within the DEM. The first layer whose top surface is greater
than the current proposed elevation of the grid cell is selected and
used along with the eikonal solver to propose a new elevation for the
given cell. Otherwise, the algorithm proceeds similarly to the fast
marching method, using the priority queue to fix the lowest
elevation cell and update its neighbors.