Ensure that visited nodes are not revisited in excesstopography3d
The current 3D approach violates the causality assumptions of the fast
marching method because the threshold slope changes based on the
current value of the elevation. This can lead to situations where
previously fixed elevation values get updated to lower values. At the
very least, this breaks the assumptions of our priority queue, namely
that the priorities of previously deleted elements are not decreased
again. We can make this work for now by storing -1 in the back pointer
of deleted elements and changing the conditions to ensure we don't
revisit a previously deleted elements.

We might also consider reinserting such nodes with insert rather than
using decrease_key.