excesstopography3d: Propose minimal consistent elevation
We define a "consistent" elevation as one whose value proposed by the
eikonal solver using the threshold slope of a given layer is below the
top of that layer in the lithstack. We choose the lowest layer that
generates a consistent elevation and update the excess topography DEM
with that elevation. This prevents the problems considered in the
previous few commits, where updates might lower the elevation below
the currently considered pixel, which then can cause updates to
propagate to previously "fixed" pixels.

This setup is quite similar to what is used in the MATLAB
implementation of TTLEM3D.