Implement excess topography (#46)
src/excesstopography.c implements excess topography computation based
on solving the eikonal equation over the gridded DEM using the fast
marching method and the fast sweeping method. The eikonal equation limits 
the maximum slope in regions of the DEM where this slope is exceeded. 
In regions of low slope, the resulting topography follows the original DEM. 
The return value (float *excess) from the excesstopography function is not the
excess topography per se, but the constrained topography. The excess
topography can be computed by subtracting the constrained topography
from the original.

Furthermore, it also implements a modification of the fast marching method
to estimate excess topography with three-dimensional lithological
variability. The lithstack array is a (nlayers x nrows x ncols) array
containing the elevations of the top surfaces of each layer at each
grid cell within the DEM. The algorithm proceeds similarly to the fast
marching method, using the priority queue to fix the lowest
elevation cell and update its neighbors. The main difference is in how 
the proposal for each neighbor is generated.

We define a "consistent" elevation as one whose value proposed by the
eikonal solver using the threshold slope of a given layer is below the
top of that layer in the lithstack. We choose the lowest layer that
generates a consistent elevation and update the excess topography DEM
with that elevation. This prevents the problems considered in the
previous few commits, where updates might lower the elevation below
the currently considered pixel, which then can cause updates to
propagate to previously "fixed" pixels.

This setup is quite similar to what is used in the MATLAB
implementation of TTLEM3D.

test/excesstopography.cpp implements property-based tests of the
excesstopography solver using a random DEM generated with the pcg4d
function used elsewhere. The test evaluates the upwind
discretization of the gradient used by the fast marching method. The
gradient should never exceed the given threshold. It is lower than the
threshold when constrained by the DEM and at the threshold otherwise.

The C standard math library is needed because we need to calculate square roots
in the eikonal solver. We link libm on Linux and other platforms that supply a
separate math library. This is checked in src/CMakeLists.txt using a method borrowed 
from An Introduction to Modern CMake by Henry Schreiner et al.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>