Add backlinks to gwdt
The `prev` array will contain the index of the previous neighbor along
the geodesic path linking the current pixel to its presill. If a null
pointer is passed in for this array, it is not written to, so it can
be treated as optional.

A test is added using the backlinks to ensure that the cost is
computed correctly.