Remove the -g flag from compile options
I added the -g to generate debug information because I had thought that
it was not getting generated properly when
CMAKE_BUILD_TYPE=Debug. This was actually because the default
CMAKE_BUILD_TYPE is not Debug but a toolchain-specific value. If I
explicitly build with CMAKE_BUILD_TYPE=Debug, -g is set and we get
debug information.