Run only one Debug job (#63)
Fixes #61

A new job is added to .github/workflows/ci.yaml that builds the Debug
configuration and runs the tests with AddressSanitizer on Ubuntu. The
other jobs build the Release configuration and run the tests on
various operating systems. The Release jobs also continue to build the
shared library to confirm that it links properly.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>