Refactor tests in random_dem.cpp (#66)
This commit refactors test/random_dem.cpp to make the tests more clear. 
Each test is extracted to its own function, and the random_dem_test 
function sets up the data, calls the appropriate libtopotoolbox 
functions, and then calls the appropriate test functions.

test/gwdt.cpp, which largely duplicated test/random_dem.cpp, is now 
merged with test/random_dem.cpp. Its tests are put into a function 
test_gwdt, which is run with the others. This improves the test speed,
because the flow routing functions like fillsinks only need to be run once.

Undefining the NDEBUG flag in `random_dem.cpp` allows us to use 
the `assert` macro in our tests in Release mode. Otherwise the asserts 
are no-ops, and the tests are useless. Since the flag only applies to 
the `random_dem` tests, it should not affect any asserts in the compiled
library.

Eventually we may want to use a more complete testing framework with
proper testing macros.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>