Run the random_dem_test with column- and row-major layouts
Since the arrays are contiguous, and we reverse the dimensions for the
row-major array, this works without any major changes to array
indexing.

This reduces the number of tests from 100 to 50 for each of the two,
but employs different RNG seeds for both of them, so that we still
test 100 different random DEMs.