Call initializers when allocating arrays in random_dem_test
This should fix the errors seen in CI on macos-latest.

Those errors occur only in the strided column major layout test. This
test is not expected to work completely for functions other than
fillsinks, because they do not access the array using the appropriate
strides. However, the array accesses with the usual column-major
strides are still valid memory reads, even if they are accessing the
incorrect elements of the array.

If these elements are initialized by the allocation with `new` to
reasonable values, then this test will still pass -- it is as if the
random elevation generator picked the initialization values for these
in-between elements. However, if there are invalid values like NaNs in
the initial values, then the test will likely fail because the
behavior of the libtopotoolbox functions in the presence of NaNs is
not really specified.

Accessing the array when uninitialized is undefined behavior, and I
guess macos ends up initialized one of the arrays with invalid
values. Using the value initializer syntax with parentheses after the
array size will initialize everything to zero, and the test will pass.

It is probably good practice to make sure these arrays are initialized
even if we do intend to change the behavior of the identifyflats and
gwdt tests so that they handle the strided access case properly.