Use doxygenfile directive to automatically generate API listing
Closes #88

The doxygenfile direction outputs an entry for everything Doxygen
documents in the given file. Since the public API is defined in
include/topotoolbox.h, this will work.

The :sections: func option displays only functions, to prevent
displaying the unhelpful TOPOTOOLBOX_VERSION_* defines. This may need
to be changed in the future if additional declarations (i.e. structs,
other defines) need to be documented.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>