Formatting changes to the docs
@bgailleton: I noticed that your documentation wasn't displaying
properly on our website. You need the `/**` at the start of the
comment for Doxygen to pick it up. I also changed some of the
indentation, but this is actually not very consistent across the
different doc comments anyway. At the moment clang-format doesn't
really format the doc comments. It might be a good idea to
find some tooling to maintain these more consistently.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>