Reformat the fillsinks documentation
The parameter description on the same line as the @param command
should be a short description of the array. This will be rendered on
the same line as the parameter name in the generated
documentation. Additional details for each parameter are placed in a
@parblock after the corresponding @param command. These should contain
a succinct description of the array type and size as well as any
important information regarding the parameter.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>