Label the target pixel of sinks/outlets correctly
Fixes #119

When a pixel has no downstream neighbors, its flow direction is set to
0 in `flow_routing_d8_carve`. The target pixel of such a pixel is
undefined. If it points to itself, then the flow accumulation of that
pixel will be doubled unless the `source==target` condition is
checked. It is more useful to set the target pixel to a sentinel
value (`-1`) and then check that `target >= 0` whenever it is used.

src/flow_routing.c implements this new behavior
test/random_dem.cpp tests that the downstream pixels are within the
array bounds and that pixels with flowdir==0 have target==-1.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>