Add a POSIX makefile to src
src/Makefile: The easiest way to incorporate libtopotoolbox into our R
bindings seems to be through a POSIX-standard Makefile. This is not
too complicated since for R we only need to build a static library out
of everything in the src/ directory. We don't need to worry about
tests or documentation. It does not support OpenMP, but we might be
able to add that later.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>