Turn warnings into errors in the docs build and fix warnings
docs/CMakeLists.txt adds the `-W --keep-going` options to
`sphinx-build`, which will error on a warning.

The warning that the `_static` directory does not exist is fixed by
creating a `_static` directory and moving the logo file into it.

The warning that ARCHITECTURE.md is not included in the toctree is
fixed by adding dev.rst to create a developer documentation table of
contents where ARCHITECTURE and CONTRIBUTING can more comfortably
reside.

.github/workflows/ci.yaml now builds the docs during the debug-build
job to alert us to any problems.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>