Refactor test/random_dem.cpp
This commit moves much of the testing infrastructure into a class
called `FlowRoutingData`. This takes care of allocating/deallocating
the data and intermediate arrays, calling the various libtopotoolbox
functions and running the test functions. This refactoring is
preparation for turning test/random_dem.cpp into a benchmark that can
be used for performance monitoring.

The arrays are all stored as std::vectors, so they are created and
destroyed automatically with `FlowRoutingData`. The constructor for
`FlowRoutingData` resizes all of these vectors to the appropriate
dimensions, then initializes the `dem` and `fraction` arrays with
random values and ones, respectively.

The `route_flow` and `route_flow_hybrid` member functions call the
various libtopotoolbox flow routing functions as before. Separating
these out will eventually allow us to benchmark the flow routing
routines by timing each of these functions whenever they are called.

Because gradients are not required for the flow routing, the
`gradient8` variants are called by their own member functions
`FlowRoutingData::gradient8` and `FlowRoutingData::gradient8_mp` for
multiprocessing. These are also amenable to future benchmarking.

The `runtests` member function replaces the `random_dem_test`
function. It runs the libtopotoolbox flow routing and gradient code
through the appropriate `FlowRoutingData` member functions then calls
the test predicates. The test predicates are currently left in the
global namespace, though they could easily -- and perhaps should -- be
made part of the class.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>