Supply boundary conditions to `fillsinks` and `fillsinks_hybrid`
The new `bc` argument to `fillsinks` and `fillsinks_hybrid` is a
`uint8_t` array of the same dimension as the DEM with a value of 1 at
each pixel that should be pinned to its original values in the DEM and
0 otherwise. This allows some additional flexibility in determining
the "boundary" of the DEM. For example, if you only wanted to fill a
smaller region within a DEM.

It also allows us to flag NaNs, so that we can reproduce the behavior
of TopoToolbox v2 with NaNs, resolving #20. Specifically, if pixel `p`
of the DEM is a NaN, then setting `dem[p] = -INFINITY` and `bc[p] =
1`, will prevent that pixel from being filled, treating it as a known
sink.

Some changes are also made to src/morphology/reconstruct.c to ensure
that the comparisons work properly if there are NaNs in the mask image
supplied to `reconstruct`. A NaN is treated as a missing value, so
that its presence does not influence its neighbors one way or the
other. Pixels will be filled according to the grayscale reconstruction
formula as if their neighborhoods do not include neighboring NaNs.

I have not included tests of the new boundary condition behavior,
though the `random_dem.cpp` tests have been adapted to use the new
`bc` argument with the standard boundaries. Property-based tests could
be added to ensure that the boundary pixels are processed
correctly. The forthcoming snapshot tests may also be useful for
verifying fillsinks when NaNs are preprocessed in different ways.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>