Implement snapshot testing framework (#136)
test/snapshot.cpp compares the output of `fillsinks` to the known
outputs in the TopoToolbox/snapshot_data repository. It is structured
similarly to test/random_dem.cpp, with a class that handles loading
all of the data and a `runtests` method that runs the libtopotoolbox
functions and the comparisons.

This adds a test-time dependency on GDAL, which is used to read the
snapshot data from disk. CMake is used to detect GDAL during
compilation. If GDAL is not detected the snapshot tests are not
compiled and will not be run by CTest.

The C++17 std::filesystem library is used for cross-platform
filesystem access, so the supported C++ standard is incremented in
CMakeLists.txt.

The TopoToolbox/snapshot_data repository is added as a submodule at
test/snapshots. This repository only contains a README and a file with
the SHA-256 checksums of all of the expected snapshot data. The data
must be downloaded separately from the releases of that repository and
extracted in `test/snapshots`.

If the test/snapshots/data repository is not present or the snapshot
data have not been extracted there, the test will pass.

The snapshot tests are run on Ubuntu in CI. Installing GDAL on macOS
and Windows is possible, but slow, so that task is left for future
work.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>