Implement drainagebasins
Fixes #118

include/topotoolbox.h contains the function prototype and documentation.

src/drainagebasins.c implements drainagebasins

The algorithm is a union-find algorithm that takes advantage of the
topological sorting of the nodes. A pixel is always upstream of any
pixel after it in the `source` list. If a `source` corresponds to a
`target == -1`, then the source has no downstream neighbors and is by
definition the root of the tree, and we label a new drainage basin
there. Otherwise each source pixel is identified with the drainage
basin of its target pixel, which should already be labeled by the type
the source pixel is reached.

src/CMakeLists.txt is updated to compile src/drainagebasins.c

test.random_dem.cpp runs `drainagebasins` and tests that each pixel
has the same basin as its downstream neighbor.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>