Implement snapshot tests for `identifyflats` and handle NaN s
test/snapshot.cpp runs `identifyflats` on the various snapshot data
sets.

src/identifyflats.c has one small change to ensure that NaNs are
handled consistently with TopoToolbox v2. Neighboring NaNs are treated
as if they lie below the current pixel, but NaNs are not themselves
flats. fminf ensures consistent NaN treatment, and the NaN equality
semantics ensures that NaNs are never set.

The sills are already handled correctly, since they are only compared
to neighbors that are known to be flats.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>