Change name of cumtrapz to streamquad_trapz
`streamquad` denotes quadrature along a stream network, and `trapz`
denotes the trapezoidal rule. One could imagine `streamquad_midp` or
`streamquad_simpson` for other quadrature rules. The 'cumulative'
aspect of the integral is not captured by the name, but it returns a
node attribute list rather than a scalar value, so it must be a
cumulative integral.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>