Output test data on failure
This refactors the snapshot tests so that each test has its own
function, which scans through the data and compares it to the
snapshot. If the two DEMs do not match, the data is written to a file
using `write_data_to_file` and an error value is returned. These are
accumulated by `runtests` and used to control output and the eventual
success or failure of the entire snapshot test suite. The test suite
runs to completion even if one of the snapshot tests fails.