Output test data when snapshot tests fail (#151)
This refactors the snapshot tests so that each test has its own
function, which scans through the data and compares it to the
snapshot. If the two DEMs do not match, the data is written to a file
using `write_data_to_file` and an error value is returned. These are
accumulated by `runtests` and used to control output and the eventual
success or failure of the entire snapshot test suite. The test suite
runs to completion even if one of the snapshot tests fails. The data are saved in the same location as the snapshot test data with a "test_" filename prefix. This may be in the CMake build folder if the test is being run with CTest as on CI.

`write_data_to_file` copies the metadata from `src_filename` and
creates a new dataset at `dst_filename` with that metadata. It then
writes `output` into the first band of the destination dataset.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>