Compute gradient8 as the steepest downward gradient
gradient8 as computed by TopoToolbox 2 and described in O'Callaghan
and Mark (1984) computes the steepest downward gradient, not the
steepest absolute gradient. This commit makes that change in
src/gradient8.c and reconfigures the random_dem test in
test/random_dem.cpp to test the property that the computed gradient is
greater than or equal to the gradients between the central pixel and
its nearest neighbors.