Change CarveMinVal to a double (#155)
MATLAB's implementation of `gwdt_computecosts` uses a double precision
floating point value for `CarveMinVal` even when the costs arrays are
`float`. This results in subtle numerical differences between the
libtopotoolbox implementation and the MATLAB implementation. This
changes libtopotoolbox so that it matches the MATLAB
implementation.

This change should be more or less inconsequential from the scientific
perspective. I do not believe, for example, that it is responsible for
the discrepancies noted in #122, though that hypothesis remains to be
confirmed. The change is made only to avoid unnecessary discrepancies
between the two implementations so that issues like #122 are easier to
resolve.

There is a very slight performance cost to using a `double` rather
than a `float` here. However, MATLAB must also deal with the same
performance cost, and it is not the bottleneck in `gwdt_computecosts`.

When `CarveMinVal` is a double, the intermediate computation is
promoted to double precision, and the result is rounded to single
precision to be stored in the `costs` array. This is not identical to
the same computation performed entirely in single precision
arithmetic, and this causes the numerical differences observed between
C and MATLAB.

The problem is exacerbated because the constant 0.1 cannot be
represented exactly as a binary floating point number, so it must be
rounded. Indeed, it seems like using a constant that can be
represented exactly in both precisions (such as 0.125) gets rid of
these numerical differences because no rounding is necessary. One
could use `float CarveMinVal = 0.125f` here and `CarveMinVal = 0.125`
in MATLAB, and one will generally obtain the same answers. However,
this seems like a fragile solution as any future changes to this
constant must be aware of the differences in its floating point
representation.

An alternative solution is to change the constant in MATLAB so that it
is the same precision as the other arrays in this expression via
something like `CarveMinVal = cast(0.1,'like',D)`. The difference
between this change and the current commit is mostly a matter of
choice. I have chosen to make the change in libtopotoolbox to avoid
changing observable behavior in the MATLAB implementation
unnecessarily.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>