Use GDALDataset::Create rather than GDALDataset::CreateCopy (#156)
We use GDAL to save the snapshots for diagnostic
purposes. GDALDataset::CreateCopy copies all of the metadata from the
source data set including the data type. In the case when the output
data has a different data type from the source dataset, the data are
converted to the type of the source dataset. This works for the
existing snapshot tests, but it does not allow us to save auxiliary
data unless they have the same type as an available source data set.

This change does not affect the test suite itself.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>