Implement TAP output from snapshot tests
This replaces the custom test output with one that conforms to the
Test Anything Protocol version 14. We don't currently use this
information because our test runner, ctest, does not do anything, but
it provides the same amount of information as we were providing
previously in a standardized format.

Related to #158