Do not assign drainage basins to invalid pixels (#161)
This PR fixes a problem with drainage basins which would lead to pixels that are not part of the flow network such as NaNs in the DEM being assigned to their own drainage basin.

The problem was that sinks, outlets and invalid pixels are all assigned placeholder edges by flow_routing_d8_carve that point from them to -1. We were using this placeholder edge to detect when to create a new drainage basin. This works for the sinks and outlets, but also creates a drainage basin for the invalid pixels. This PR switches this behavior to 1. skip these placeholder edges and 2. test whether the target pixel has not yet been assigned to a drainage basin. This matches the behavior of the MATLAB implementation, leaving invalid pixels with a 0 label for their drainage basin.