Switch stream traversal functions to use 0-based arrays
This unifies the stream network traversal interface with the flow
network. I would like it to be possible to pass source and target
arrays from either a FLOWobj/FlowObject or a STREAMobj/StreamObject to
our various traversal functions without needing to worry about which
indexing convention is used. 0-based arrays make the most sense for C,
so we'll assume that source and target are 0-based indices and
decrement in MATLAB.