Add bitwise or accumulation to traverse_up_u32_or_and
As described in #164, this makes the semiring pattern in the traversal
more obvious.

A test is added to test/random_dem.cpp that computes drainage basins
using traverse_up_u32_or_and and ensures that it is identical to our
existing drainagebasins computation. This also tries a new testing
pattern where test-specific data is initialized within the test
function rather than as part of the very large FlowRoutingData class.

The test currently accesses the traverse_basins array out of bounds
when the source pixel is a sink, because we construct the placeholder
sink->-1 edges in the flow network.