Download snapshot data using FetchContent (#176)
This replaces the curl script in the CI configuration with a CMake
FetchContent declaration. If snapshot data is present in the
test/snapshots/data directory during configuration, then it will be
used instead of downloading the new snapshot data.

This change increases the duplication of the snapshot version number,
which now must be changed in .github/workflows/ci.yaml to make sure
that the correct data is cached, and in test/CMakeLists.txt, to make
sure the correct data is fetched. The hash of the snapshot archive is
also needed by FetchContent, and this is different from the checksums
that we provide in the snapshot_data repository.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>