Implement hillshade
This patch implements `hillshade` for computing hillshade
visualizations of digital elevation models. It successfully reproduces
the outputs of `hillshade` from TopoToolbox v2 on the snapshot test
data.

Two new functions are added to the API. `gradient_secondorder`
computes the gradient of the DEM using a second-order finite
difference approximation that is equivalent to that used by MATLAB in
`surfnorm` and by `numpy.gradient`. This function is used elsewhere by
TopoToolbox functions (such as `aspect`) and might be useful in those
places, though its performance against optimized versions has not been
thoroughly evaluated.

`hillshade` computes the hillshade of the DEM by computing normal
vectors using `gradient_secondorder` and evaluating the dot product of
the normal vectors and the specified lighting vector.

A snapshot test of the hillshade raster using v1.6.0 of the snapshot
data is added to `test/snapshot.cpp`, and the snapshot data version is
bumped in the necessary locations.