Clearly specify the grid-centered coordinate system in hillshade
This choice of coordinate system is arbitrary, but it makes the most
sense to work entirely within a grid-based coordinate system, where
the "x" axis increases with the first dimension of the grid and the
"y" axis increases with the second dimension of the grid. This uses no
external information on the orientation of the image. However, the
relationship between grid coordinates and geographic coordinates
depends on both the coordinate system of the image and the memory
layout of the grid.

GDAL's default coordinate system has the first dimension increasing
from the left to the right as one looks at the image, while the second
dimension increases from top to bottom. If this image is oriented so
that north is up, this means that the azimuth desired by `hillshade`
increases from the east (0 degrees) towards the south (90 degrees).

If the array is transposed so that it is now in "column-major"
order (the default in MATLAB), then the azimuth angle will increase
from the south (0 degrees) towards the east (90 degrees).

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>