Modify offsets to make row- and column-major indexing consistent  (#186)
Flow routing should produce the same results for identical DEMs stored
in either row- or column-major order. However, we search sequentially
through neighboring pixels to find the steepest gradient, and we
choose the first neighbor we encounter if two neighbors have the same
gradient. For column-major arrays, we start with the east neighbor and
check in a clockwise direction. But for row-major arrays, using the
same offsets in the i and j directions starts with the south neighbor
and checks counterclockwise.

This change introduces a new parameter (order) that is zero for
column-major arrays and one for row-major arrays. Think about it as
the index of the dimension whose index changes the fastest as you
iterate linearly through the array. This parameter is used to
construct the offset arrays so that neighbors are visited in the same
order regardless of the memory layout of the array.

This is a breaking API change to libtopotoolbox, and downstream consumers will need to be changed. To preserve the current behavior, pass 0 to the order parameter of flow_routing_d8_carve and flow_routing_d8_edgelist. As noted above, this behavior is incorrect. You should pass 0 for column-major arrays and 1 for row-major arrays.