Add a lower-memory version of hillshade (#192)
`hillshade_fused` should behave identically to `hillshade`, but it
requires no intermediate arrays. It is slightly slower than
`hillshade` because of different memory access patterns when computing
the gradients and differences in handling boundary conditions. It
seems like the performance gap closes for larger DEMs, which have more
of their pixels on the interior. One might choose to use
`hillshade_fused` for larger DEMs.

`hillshade` is also modified so that it takes only two intermediate
arrays. These are used to store the surface gradients. The normal
vector computation is now inlined.

A snapshot test is added to ensure that `hillshade_fused` also
replicates the behavior of TT2's hillshade.

This is a breaking change. The change to the `hillshade` API will need
to be patched in downstream libraries.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>