Implement `lowerenv` (#196)
This function computes the lower convex envelope of a stream profile and
is used in the knickpointfinder to model a convex stream profile,
deviations from which indicate the presence of knickpoints.

This more or less copies the MATLAB implementation with a few
changes:

The minimum gradient for each reach is computed by explicitly
accumulating the minimum over the network rather than by sorting. This
avoids dynamic memory allocation for the sorting, but does require
that we loop over the whole stream network even when we are examining
a portion of it.

`allpred` is inlined into the function, since it is only used
once. This is noted with a few comments.

`allpred` is modified so that it represents an upstream traversal in
the Boolean semiring. This is different from the MATLAB
implementation, which does not have a semiring update step. The MATLAB
implementation can miss some reaches above knickpoints and produces
stream networks that are not convex.

A random_dem test is added to check that the resulting profile is
convex except at knickpoints. Knickpoints are "randomly" generated
along the stream profile. The test checks that the gradient of each
edge flowing into a node is greater than the gradient leaving that
node, except at knickpoints, where this will not be true.

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>